Add windows system integration

+++ src/PboViewer GUI/Windows/PboViewerMainWindow.xaml.cs
@@ -1,8 +1,10 @@
 ﻿using Avalonia;
 using Avalonia.Controls;
+using Avalonia.Input;
 using Avalonia.Markup.Xaml;
 using PboViewer.ViewModels;
 using PboViewer.Views;
+using System.Diagnostics;
 using System.IO;
 
 namespace PboViewer.Windows
@@ -12,7 +14,7 @@ namespace PboViewer.Windows
     /// </summary>
     public class PboViewerMainWindow : FluentWindow
     {
-        public TextBlock Title;
+        new public TextBlock Title;
 
         public PboViewerMainWindow()
         {
@@ -25,14 +27,20 @@ namespace PboViewer.Windows
         private void InitializeComponent()
         {
             AvaloniaXamlLoader.Load(this);
-            Closing += PboViewerMainWindow_Closing;
-            Title = this.FindControl<TextBlock>("Title");
 
+            // Register general event
+            Closing += PboViewerMainWindow_Closing;
+
+            // Set properties
+            Title = this.FindControl<TextBlock>("Title");
             DataContext = new PboViewerMainWindowViewModel();
         }
 
         private void PboViewerMainWindow_Closing(object sender, System.ComponentModel.CancelEventArgs e)
         {
+            if (((PboViewerMainWindowViewModel)DataContext).Child is not EditorView)
+                return;
+
             EditorView editorView = ((EditorView)((PboViewerMainWindowViewModel)DataContext).Child);
             if (editorView.ToDelete)
                 Directory.Delete(editorView.EditorPath, true);
